Support dry run  class execution #59

+++ bin/worker.execute.py
@@ -380,11 +380,12 @@ class WorkerExecution:
             required_argument_count += 1
             self.docker_file += self.applyTemplate(JAVA_CLASS_EXECUTION_TEMPLATE)
             self.docker_file += self.docker_file.replace('$CLASS_FOR_EXECUTION', self.config.class_for_execution)
-            self.deployable_jars = Path(Path.home().joinpath('.local/state/' + self.config.name + '/repos/public/net.splitcells.network/target/' + self.config.name + '/deployable-jars/'))
-            self.deployable_jars.mkdir(parents=True, exist_ok=True)
-            shutil.copytree(str(Path.home().joinpath('.local/state/' + self.config.name + "/repos/public/" + self.config.name + "/target/deployable-jars/"))
-                , str(self.deployable_jars)
-                , dirs_exist_ok=True)
+            if not self.config.dry_run:
+                self.deployable_jars = Path(Path.home().joinpath('.local/state/' + self.config.name + '/repos/public/net.splitcells.network/target/' + self.config.name + '/deployable-jars/'))
+                self.deployable_jars.mkdir(parents=True, exist_ok=True)
+                shutil.copytree(str(Path.home().joinpath('.local/state/' + self.config.name + "/repos/public/" + self.config.name + "/target/deployable-jars/"))
+                    , str(self.deployable_jars)
+                    , dirs_exist_ok=True)
         if required_argument_count == 0:
             raise Exception("Either `--command`, `--executable-path` or `--class-for-execution` needs to be set:" + str(self.config))
         if required_argument_count > 1:
